refactor(oxlint/lsp): inline `IsolatedLintHandler` (#18492)
Refactors the oxlint LSP linting flow by removing `IsolatedLintHandler` and inlining its responsibilities into `ServerLinter`.
The `IsolatedLintHandlerFileSystem` was renamed and put into an extra file.

This will help us to align the CLI with LSP config/runner building
