+++ apps/oxlint/src/lsp/mod.rs
@@ -2,7 +2,7 @@ mod code_actions;
 mod commands;
 mod config_walker;
 mod error_with_position;
-mod isolated_lint_handler;
+mod lsp_file_system;
 mod options;
 mod server_linter;
 #[cfg(test)]
